Fix.Search util분리, 헤더변경
