+++ src/scss/components/_button.scss
@@ -220,3 +220,7 @@
 .nosearch-btn {
   cursor: not-allowed;
 }
+
+.cancel-btn {
+  cursor: not-allowed;
+}
