+++ src/util/dateUtil.js
@@ -1,13 +1,3 @@
-export function commaNumber(number) {
-  if (number !== null) {
-    return (typeof number === "string" ? number : number.toString()).replace(
-      /\B(?=(\d{3})+(?!\d))/g,
-      ","
-    );
-  }
-  return 0;
-}
-
 // 2025-01-01을 n월n일로 리턴
 export function monthDateFormat(dateString) {
   if (dateString) {
@@ -50,18 +40,4 @@ export const getDatesInRange = (startDate, endDate) => {
   }
 
   return dates;
-};
-
-// 모달 여는함수
-export const handleOpenModal = (currentModal) => {
-  if (currentModal.current) {
-    currentModal.current.showModal();
-  }
-};
-
-// 모달 취소 버튼 클릭 : 창 닫기
-export const handleCancelModal = (currentModal) => {
-  if (currentModal.current) {
-    currentModal.current.close();
-  }
 };
